* dev: (29 commits)
  [Feat] PR 수정 사항 수정
  [Feat] 꿀팁 모달 스타일 조정
  [Fix] d-day 컬러 정정 및 기업상세 이미지 fit 수정
  [Feat] 공고디테일 상시채용 마지막 시간 제거
  [Fix] jds 타입 변경
  [Fix] 변경된 타입 맞게 지정
  [Feat] 취업꿀팁 스타일 조정중 임시커밋
  [Feat] 댓글 기반 내용 전부 수정
  [Feat] posting 스타일 조정
  [fix] OU 빌드테스트 실패 해결 시도
  [Fix] OU 전부 업데이트 완료
  [Feat] posting card 스타일 정정
  [Feat] 이벤트페이지 틀생성
  [Feat] 404/500 페이지 스타일 조정
  [Feat] 기업 디테일 스타일 정리
  [Fix] 스펙리스트 페이지 회원가입 미노출 수정
  [Feat] 모달백그라운드 홈으로 가기 추가
  [Feat] jd 디테일 스타일 정리
  [Feat] jd 만료공고 스타일 정리
  [Feat] jd list page 스타일 정리
  ...

# Conflicts:
#	apps/gocho_web/public/images/global/MainCarousel/banner01_check.png
#	apps/gocho_web/src/component/card/communityPosting/style.ts
#	apps/gocho_web/src/component/card/jobCard/index.tsx
#	apps/gocho_web/src/component/modal/changePostingModal/style.ts
#	apps/gocho_web/src/component/modal/postingModal/component/comment/style.ts
#	apps/gocho_web/src/component/modal/postingModal/component/reComment/style.ts
#	apps/gocho_web/src/component/modal/postingModal/style.ts
#	apps/gocho_web/src/component/modal/writePostingModal/style.ts
#	apps/gocho_web/src/pages/community/postings/component/postingCard/style.ts
#	apps/gocho_web/src/pages/community/tips/component/tipCard/style.ts
#	apps/gocho_web/src/pages/companies/part/headerPart/index.tsx
#	apps/gocho_web/src/pages/datalab/spec/component/specCard/index.tsx
#	apps/gocho_web/src/pages/datalab/spec/component/specCard/style.ts
#	apps/gocho_web/src/pages/indexComponent/part/jobPart/component/jobSmall/index.tsx
#	apps/gocho_web/src/pages/indexComponent/part/jobPart/component/jobSmall/style.ts
#	apps/gocho_web/src/pages/indexComponent/part/mainCarouselPart/constant.ts
#	apps/gocho_web/src/pages/indexComponent/part/specBestPart/component/bestUserInfo/style.ts
#	apps/gocho_web/src/pages/indexComponent/part/tipPart/component/latestCard/style.ts
#	apps/gocho_web/src/pages/indexComponent/part/tipPart/component/tipDisplay/style.ts
#	apps/gocho_web/src/pages/jd/detail/part/asidePart/index.tsx
#	apps/gocho_web/src/pages/jd/detail/part/headerPart/component/headerFix/style.ts
#	apps/gocho_web/src/pages/jd/explist/component/expJobCard/index.tsx
#	apps/gocho_web/src/pages/jd/explist/component/expJobCard/style.ts
#	apps/gocho_web/src/pages/jd/list/component/jobAdCard/index.tsx
#	apps/gocho_web/src/pages/jd/list/part/adPart/index.tsx
#	apps/gocho_web/src/pages/mypage/part/bookmarkPart/jobCard/style.ts
#	apps/gocho_web/src/pages/mypage/part/calendarPart/component/bookmark/style.ts
#	pnpm-lock.yaml
#	shared/api/bookmark/type/bookmark.ts
#	shared/api/bookmark/useUserJobBookmarkArr/util.ts
#	shared/image/global/common/mainBanner_check.png
#	shared/image/global/common/yellow_check.png
#	shared/style/package.json
#	shared/ui/common/companyCard/index.tsx
#	shared/ui/package.json

+++ shared/ui/common/companyCard/index.tsx
@@ -3,11 +3,10 @@ import Image from "next/image";
 import Link from "next/link";
 import { BsFillBookmarkFill } from "react-icons/bs";
 
-import { useAddUserBookmark, useDeleteUserBookmark } from "shared-api/bookmark";
+import { useAddCompanyBookmarkArr, useDeleteCompanyBookmarkArr } from "shared-api/bookmark";
+import { SkeletonBox } from "shared-ui/common/atom/skeletonBox";
 import { COMPANY_DETAIL_URL } from "shared-constant/internalURL";
 import defaultCompanyLogo from "shared-image/global/common/default_company_logo.svg";
-
-import { SkeletonBox } from "../atom/skeletonBox";
 
 import { CompanyCardProps, CompanyCardSkeleton } from "./type";
 import {
@@ -27,12 +26,19 @@ export const CompanyCard: FunctionComponent<CompanyCardProps | CompanyCardSkelet
   isBookmarked,
   userId,
   isSkeleton,
-  refetchUserBookmark,
 }) => {
   const [imageSrc, setImageSrc] = useState(companyData?.logoUrl as string);
 
-  const { mutate: addMutate } = useAddUserBookmark();
-  const { mutate: deleteMutate } = useDeleteUserBookmark();
+  const { mutate: addMutate } = useAddCompanyBookmarkArr({
+    id: companyData?.id as number,
+    logo_url: companyData?.logoUrl as string,
+    name: companyData?.name as string,
+  });
+  const { mutate: deleteMutate } = useDeleteCompanyBookmarkArr({
+    id: companyData?.id as number,
+    logo_url: companyData?.logoUrl as string,
+    name: companyData?.name as string,
+  });
 
   const size = "";
   const sector = "";
@@ -46,31 +52,11 @@ export const CompanyCard: FunctionComponent<CompanyCardProps | CompanyCardSkelet
   }
 
   const addCompanyBookmark = () => {
-    return (
-      userId &&
-      addMutate(
-        { userId, likeType: "company-bookmarks", elemId: companyData.id },
-        {
-          onSuccess: () => {
-            refetchUserBookmark();
-          },
-        }
-      )
-    );
+    if (userId) addMutate({ userId, elemId: companyData.id });
   };
 
   const deleteCompanyBookmark = () => {
-    return (
-      userId &&
-      deleteMutate(
-        { userId, likeType: "company-bookmarks", elemId: companyData.id },
-        {
-          onSuccess: () => {
-            refetchUserBookmark();
-          },
-        }
-      )
-    );
+    if (userId) deleteMutate({ userId, elemId: companyData.id });
   };
 
   return (
